symmetry to soslam.cpp with feature points

+++ include/SymmetryFactor.h
@@ -19,15 +19,18 @@
 #include <gtsam/nonlinear/NonlinearFactorGraph.h>
 #include <opencv2/opencv.hpp>
 
-namespace gtsam_soslam {
+namespace gtsam_soslam
+{
     /**
      * SymmetryFactor
      * factor between Pose3 and ConstrainedDualQuadric
      */
     class SymmetryFactor
-            : public gtsam::NoiseModelFactor2<gtsam::Pose3, ConstrainedDualQuadric> {
+        : public gtsam::NoiseModelFactor2<gtsam::Pose3, ConstrainedDualQuadric>
+    {
     public:
-        enum MeasurementModel {
+        enum MeasurementModel
+        {
             STANDARD,
             TRUNCATED
         };
@@ -37,9 +40,9 @@ namespace gtsam_soslam {
         AlignedBox2 measured_;
         std::string label_;
         boost::shared_ptr<gtsam::Cal3_S2> calibration_; //< camera calibration
-        typedef NoiseModelFactor2 <gtsam::Pose3, ConstrainedDualQuadric> Base;
+        typedef NoiseModelFactor2<gtsam::Pose3, ConstrainedDualQuadric> Base;
         MeasurementModel measurementModel_;
-        std::vector<std::vector<std::pair<double, double>>> nearest_edge_point_;
+        std::vector<std::vector<std::pair<int, int>>> nearest_edge_point_;
         int sigma_scc_;
 
     public:
@@ -47,7 +50,7 @@ namespace gtsam_soslam {
 
         /** Default constructor */
         SymmetryFactor()
-                : label_("None"), measurementModel_(STANDARD), sigma_scc_(1) {};
+            : label_("None"), measurementModel_(STANDARD), sigma_scc_(1){};
 
         SymmetryFactor(const AlignedBox2 &measured,
                        const cv::Mat &image,
@@ -55,18 +58,18 @@ namespace gtsam_soslam {
                        const boost::shared_ptr<gtsam::Cal3_S2> &calibration,
                        const gtsam::Key &poseKey, const gtsam::Key &quadricKey,
                        const gtsam::SharedNoiseModel &model,
-                       const std::vector<std::vector<std::pair<double, double>>> &nearest_edge_point,
+                       const std::vector<std::vector<std::pair<int, int>>> &nearest_edge_point,
                        const MeasurementModel &errorType = STANDARD,
 
                        const int &sigma_scc = 1)
-                : Base(model, poseKey, quadricKey),
-                  image_(image),
-                  measured_(measured),
-                  label_(label),
-                  calibration_(calibration),
-                  measurementModel_(errorType),
-                  nearest_edge_point_(nearest_edge_point),
-                  sigma_scc_(sigma_scc) {};
+            : Base(model, poseKey, quadricKey),
+              image_(image),
+              measured_(measured),
+              label_(label),
+              calibration_(calibration),
+              measurementModel_(errorType),
+              nearest_edge_point_(nearest_edge_point),
+              sigma_scc_(sigma_scc){};
 
         SymmetryFactor(const AlignedBox2 &measured,
                        const cv::Mat &image,
@@ -75,23 +78,29 @@ namespace gtsam_soslam {
                        const gtsam::Key &poseKey, const gtsam::Key &quadricKey,
                        const gtsam::SharedNoiseModel &model,
                        const std::string &errorString,
-                       const std::vector<std::vector<std::pair<double, double>>> &nearest_edge_point,
+                       const std::vector<std::vector<std::pair<int, int>>> &nearest_edge_point,
                        const int &sigma_scc = 1)
-                : Base(model, poseKey, quadricKey),
-                  image_(image),
-                  measured_(measured),
-                  label_(label),
-                  calibration_(calibration),
-                  nearest_edge_point_(nearest_edge_point),
-                  sigma_scc_(sigma_scc) {
-            if (errorString == "STANDARD") {
+            : Base(model, poseKey, quadricKey),
+              image_(image),
+              measured_(measured),
+              label_(label),
+              calibration_(calibration),
+              nearest_edge_point_(nearest_edge_point),
+              sigma_scc_(sigma_scc)
+        {
+            if (errorString == "STANDARD")
+            {
                 measurementModel_ = STANDARD;
-            } else if (errorString == "TRUNCATED") {
+            }
+            else if (errorString == "TRUNCATED")
+            {
                 measurementModel_ = TRUNCATED;
-            } else {
+            }
+            else
+            {
                 throw std::logic_error(
-                        "The error type \"" + errorString +
-                        "\" is not a valid option for initializing a SymmetryFactor");
+                    "The error type \"" + errorString +
+                    "\" is not a valid option for initializing a SymmetryFactor");
             }
         }
 
@@ -107,9 +116,9 @@ namespace gtsam_soslam {
          * @param H2 the derivative of the error wrt quadric (1x9)
          */
         gtsam::Vector evaluateError(
-                const gtsam::Pose3 &pose, const ConstrainedDualQuadric &quadric,
-                boost::optional<gtsam::Matrix &> H1 = boost::none,
-                boost::optional<gtsam::Matrix &> H2 = boost::none) const override;
+            const gtsam::Pose3 &pose, const ConstrainedDualQuadric &quadric,
+            boost::optional<gtsam::Matrix &> H1 = boost::none,
+            boost::optional<gtsam::Matrix &> H2 = boost::none) const override;
 
         /** Evaluates the derivative of the error wrt pose */
         gtsam::Matrix evaluateH1(const gtsam::Pose3 &pose,
@@ -128,7 +137,7 @@ namespace gtsam_soslam {
         /** Prints the SymmetryFactor with optional string */
         void print(const std::string &s = "",
                    const gtsam::KeyFormatter &keyFormatter =
-                   gtsam::DefaultKeyFormatter) const override;
+                       gtsam::DefaultKeyFormatter) const override;
 
         /** Returns true if equal keys, measurement, noisemodel and calibration */
         bool equals(const SymmetryFactor &other, double tol = 1e-9) const;
@@ -137,8 +146,9 @@ namespace gtsam_soslam {
 } // namespace gtsam_soslam
 
 // Add to testable group
-template<>
+template <>
 struct gtsam::traits<gtsam_soslam::SymmetryFactor>
-        : public gtsam::Testable<gtsam_soslam::SymmetryFactor> {
+    : public gtsam::Testable<gtsam_soslam::SymmetryFactor>
+{
 };
 #endif // GTSAM_SOSLAM_SYMMETRYFACTOR_H
