+++ include/SystemState.h
@@ -24,28 +24,36 @@
 #include <gtsam/nonlinear/Values.h>
 #include <gtsam/geometry/Cal3.h>
 
-namespace gtsam_soslam {
+namespace gtsam_soslam
+{
 
-    class Detection {
+    class Detection
+    {
     public:
         std::string label;
         gtsam::Vector4 bounds;
         gtsam::Key pose_key;
         gtsam::Key quadric_key;
 
-        bool operator==(const Detection &other) const {
+        bool operator==(const Detection &other) const
+        {
             if (label == other.label && bounds == other.bounds && pose_key == other.pose_key &&
-                quadric_key == other.quadric_key) {
+                quadric_key == other.quadric_key)
+            {
                 return true;
-            } else {
+            }
+            else
+            {
                 return false;
             }
         }
 
         Detection(const Detection &other) = default;
 
-        Detection &operator=(const Detection &other) {
-            if (this != &other) {
+        Detection &operator=(const Detection &other)
+        {
+            if (this != &other)
+            {
                 label = other.label;
                 bounds << other.bounds[0], other.bounds[1], other.bounds[2], other.bounds[3];
                 pose_key = other.pose_key;
@@ -64,7 +72,8 @@ namespace gtsam_soslam {
     class StepState
     {
     private:
-        bool image_ready_=false;
+        bool image_ready_ = false;
+
     public:
         int i;
         gtsam::Key pose_key;
@@ -73,15 +82,18 @@ namespace gtsam_soslam {
         gtsam::Pose3 odom;
         std::vector<Detection> detections;
         std::vector<Detection> new_associated;
-        std::vector<std::vector<std::pair<double, double>>> nearest_edge_point;
+        std::vector<std::vector<std::pair<int, int>>> nearest_edge_point;
 
-        explicit StepState(int i = 0) : i(i), pose_key(gtsam::Symbol('x', i)) {
+        explicit StepState(int i = 0) : i(i), pose_key(gtsam::Symbol('x', i))
+        {
         }
 
         StepState(const StepState &other) = default;
 
-        StepState &operator=(const StepState &other) {
-            if (this != &other) {
+        StepState &operator=(const StepState &other)
+        {
+            if (this != &other)
+            {
                 i = other.i;
                 pose_key = other.pose_key;
                 rgb = other.rgb;
@@ -92,16 +104,17 @@ namespace gtsam_soslam {
             }
             return *this;
         }
-        void imageprepared(){ image_ready_=true;}
-        bool needSave() const{return image_ready_;}
-        void imageSaved(){ image_ready_=false;}
+        void imageprepared() { image_ready_ = true; }
+        bool needSave() const { return image_ready_; }
+        void imageSaved() { image_ready_ = false; }
         bool isValid() const
         {
             return i != 0;
         }
     };
 
-    class SoSlamState {
+    class SoSlamState
+    {
     public:
         gtsam::Pose3 initial_pose_;
         gtsam::Matrix noise_prior_;
@@ -121,10 +134,10 @@ namespace gtsam_soslam {
         StepState prev_step;
         StepState this_step;
 
-
         explicit SoSlamState(const gtsam::Pose3 &initial_pose = Constants::POSES[0], const bool &optimizer_batch = true)
-                : initial_pose_(initial_pose),
-                  optimizer_batch_(optimizer_batch) {
+            : initial_pose_(initial_pose),
+              optimizer_batch_(optimizer_batch)
+        {
 
             isam_params_.relinearizeThreshold = 0.05;
             isam_params_.relinearizeSkip = 10;
